Integrated Dynamics!
Created dynamics integrator for the project. Results are a nice heliocentric path of the satellite trailing the earth. I've been thinking about it, and I may code up the same thing but in the earth centered frame, just to see what it looks like. The orbit path given in the project is earth centered, so It may be a better sanity check. Plus if all our measurements are earth centered, it could reduce the need for an additional rotation matrix going from EMO to EME to ECEF etc.